% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{background_inputprocess}
\alias{background_inputprocess}
\title{Imports and processes the mutation count data}
\usage{
background_inputprocess(datapath, numbase)
}
\description{
Imports and processes the mutation count data
}
