% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Func.R
\name{complexconv_create}
\alias{complexconv_create}
\title{Constructs the sparse approximation of h(T_i * F_i) <- check this is true}
\usage{
complexconv_create(N, K, numbase, feat, mid)
}
\description{
Constructs the sparse approximation of h(T_i * F_i) <- check this is true
}
