% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{conv_create}
\alias{conv_create}
\title{Constructs the sparse approximation of h(T_i * F_i)}
\usage{
conv_create(N, K, numbase, feat)
}
\description{
Constructs the sparse approximation of h(T_i * F_i)
}
