% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{fragbase_indexer}
\alias{fragbase_indexer}
\title{Stores the indices of the fragment types according to which base is present
at a particular base position}
\usage{
fragbase_indexer(numbase, N)
}
\description{
Stores the indices of the fragment types according to which base is present
at a particular base position
}
