% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_Shallow.R
\name{icgc_curate}
\alias{icgc_curate}
\title{Sorts and removes duplicate entries in a mutation file output
by \code{icgc2mut()}}
\usage{
icgc_curate(mut_file, remove.nonSNP = TRUE)
}
\arguments{
\item{mut_file}{A mutation file output
by \link[=icgc2mut]{icgc2mut()}.}

\item{remove.nonSNP}{A boolean variable indicating whether the function
should remove non-single nucleotide changes. This is set by default
to \code{TRUE} as our downstream filtering method only handles single
nucleotide changes.}
}
\value{
A sorted mutation file with duplicate entries removed and, depending
on user specification, non-single nucleotide changes removed.
}
\description{
Sorts and removes duplicate entries in a mutation file output
by \code{icgc2mut()}
}
\section{Details}{

Your input mutation file must at least have the following column headers:
icgc_sample_id (i.e. the ICGC sample ID), chromosome (i.e. the Chromosome ID),
chromosome_start (i.e. the chromosome
start position), chromosome_end (i.e. the chromosome end position),
mutated_from_allele (i.e. the reference allele), and mutated_to_allele
(i.e. alternate allele).
}

\examples{
res <- icgc_curate(mutation_data)

}
