% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_Shallow.R
\name{icgc_sort}
\alias{icgc_sort}
\title{#' Loads an example ICGC file
#'
#' @return A valid ICGC input file (that can for example be used for \link[=icgc2mut]{icgc2mut()}).
#'
#' @examples
#' loadICGCexample()
#'
#' @export
loadICGCexample <- function() {Import example input dataset

load("data/example_mutation_dataset.Rda", envir = parent.frame())
}}
\usage{
icgc_sort(mut_file)
}
\description{
#' Loads an example ICGC file
#'
#' @return A valid ICGC input file (that can for example be used for \link[=icgc2mut]{icgc2mut()}).
#'
#' @examples
#' loadICGCexample()
#'
#' @export
loadICGCexample <- function() {Import example input dataset

load("data/example_mutation_dataset.Rda", envir = parent.frame())
}
}
