% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_Shallow.R
\name{loadICGCexample}
\alias{loadICGCexample}
\title{Loads an example ICGC file}
\usage{
loadICGCexample()
}
\value{
A valid ICGC input file (that can for example be used for \link[=icgc2mut]{icgc2mut()}).
}
\description{
Loads an example ICGC file
}
\examples{
loadICGCexample()

}
