% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMu_prepare.R
\name{mut_process3}
\alias{mut_process3}
\title{A function that treats the mutation input file and reorders its columns}
\usage{
mut_process3(datapath)
}
\description{
A function that treats the mutation input file and reorders its columns
}
