% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMu_prepare.R
\name{mut_process35}
\alias{mut_process35}
\title{A function that treats the mutation input file and reorders its columns}
\usage{
mut_process35(datapath)
}
\description{
A function that treats the mutation input file and reorders its columns
}
\section{Details}{

This is an auxiliary function.
}

