% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{mutation_inputprocess}
\alias{mutation_inputprocess}
\title{Imports and processes the mutation count data}
\usage{
mutation_inputprocess(datapath, numbase)
}
\description{
Imports and processes the mutation count data
}
