% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_VCF.R
\name{read_vcf}
\alias{read_vcf}
\title{Read vcf files}
\usage{
read_vcf(vcf_file, well_id = NULL, region = NULL, drop_col = FALSE)
}
\arguments{
\item{drop_col}{}
}
\description{
Read vcf files
}
\examples{

}
