% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMu_prepare.R
\name{readfast}
\alias{readfast}
\title{Read in the reference genome fasta file}
\usage{
readfast(datapath)
}
\description{
Read in the reference genome fasta file
}
\section{Details}{

This is an auxiliary function.
}

