% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{relu_bigboy}
\alias{relu_bigboy}
\title{Performs the ReLU transform on the mutational data}
\usage{
relu_bigboy(mut_path, bg_path, numbase)
}
\description{
Performs the ReLU transform on the mutational data
}
