% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{relu_transform}
\alias{relu_transform}
\title{Performs the ReLU transform on the mutational data}
\usage{
relu_transform(mut_path, bg_path, five = FALSE)
}
\description{
Performs the ReLU transform on the mutational data
}
