% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{tencode}
\alias{tencode}
\title{Returns a matrix with all possible unmutated fragment types one-hot encoded}
\usage{
tencode(numbase, N)
}
\description{
Returns a matrix with all possible unmutated fragment types one-hot encoded
}
