% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_VCF.R
\name{vcf2mut}
\alias{vcf2mut}
\title{Read vcf files and processes them / User correct assembly}
\usage{
vcf2mut(datapath)
}
\arguments{
\item{datapath}{}
}
\description{
Read vcf files and processes them / User correct assembly
}
\examples{

}
