% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_Shallow.R
\name{icgc_curate}
\alias{icgc_curate}
\title{Sorts and removes duplicate entries in a mutation file output
by \code{icgc2mut()}}
\usage{
icgc_curate(mut_file, remove.nonSNP = TRUE)
}
\arguments{
\item{mut_file}{A string or a variable referencing a mutation file output
by \link[=icgc2mut]{icgc2mut()}. You can specify a file you created, however
you have to make sure that it has the correct format (i.e. please view
\emph{Details}). Your input should either be a \code{data.frame}, a
\code{matrix}, or a \code{data.table}.}

\item{remove.nonSNP}{A boolean variable indicating whether the function
should remove non-single nucleotide changes. This is set by default
to \code{TRUE} as our downstream filtering method only handles single
nucleotide changes. You can toggle this parameter to \code{FALSE} for your
own purposes, although it should remain \code{TRUE} if you want to utilize
our entire pipeline for your analysis.}
}
\value{
A sorted mutation file with duplicate entries removed and, depending
on user specification, non-single nucleotide changes removed.
}
\description{
Sorts and removes duplicate entries in a mutation file output
by \code{icgc2mut()}
}
\section{Details}{

Your input mutation file must at least have the following column headers:
icgc_sample_id (i.e. the ICGC sample ID), chromosome (i.e. the Chromosome ID),
chromosome_start (i.e. the chromosome
start position), chromosome_end (i.e. the chromosome end position),
mutated_from_allele (i.e. the reference allele), and mutated_to_allele
(i.e. alternate allele).
}

\examples{
res <- icgc_curate(mutation_data)

}
