% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shallow_ReLU.R
\name{linreg_solver}
\alias{linreg_solver}
\title{Solves the |T * Feature - theta|^2 = 0 linear regression problem}
\usage{
linreg_solver(T, theta)
}
\description{
Solves the |T * Feature - theta|^2 = 0 linear regression problem
}
