% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_Shallow.R
\name{loadICGCexample}
\alias{loadICGCexample}
\title{Loads an example ICGC file}
\usage{
loadICGCexample()
}
\value{
A valid ICGC input file (e.g. for \code{ICGC2Mut()}).
}
\description{
Loads an example ICGC file
}
\examples{
example_data <- loadICGCexample()

}
