% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rev_EMu_prepare.R
\name{mut_count_fast}
\alias{mut_count_fast}
\title{Function Doc Pending}
\usage{
mut_count_fast(genome_path, mut_file, numbase, nb_chrom)
}
\description{
Function Doc Pending
}
\section{Details}{

Pending
}

