% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_VCF.R
\name{vcf2mut}
\alias{vcf2mut}
\title{Read vcf files and processes them / User correct assembly}
\usage{
vcf2mut(datapath, geno = "GT", assembly, five = FALSE)
}
\arguments{
\item{datapath}{The path leading to your multisampled vcf file.}

\item{geno}{An ID Format string in your vcf file indicating the index of the
genotype information.}

\item{assembly}{The path leading to your assembly file (.fa or .fa.gz).}

\item{five}{A boolean variable. A value of \code{TRUE} will lead to the function
scanning the input files for 5 bases mutation signatures as opposed to 3 bases
signatures. a value of \code{FALSE} causes the function to scan for 3 bases signatures.}
}
\value{
A background mutation signatures vector (\code{wt}), which provides
the frequency of each possible signature given an assembly file. A matrix (\code{mut_mat})
containing the mutational rate of each signature for each sample in your supplied mutation
input file.
}
\description{
Read vcf files and processes them / User correct assembly
}
\examples{
assembly <- "Homo_sapiens.GRCh37.dna.primary_assembly.fa"
mut_file <- "multisampled_mutations.vcf"

mut_sign <- vcf2mut(mut_file, "GT", assembly)

}
