% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_VCF.R
\name{vcf2mut}
\alias{vcf2mut}
\title{Read vcf files and processes them / User correct assembly}
\usage{
vcf2mut(datapath, geno = "GT")
}
\arguments{
\item{datapath}{The path leading to your vcf file.}

\item{geno}{An ID Format string in your vcf file indicating the index of the
genotype information.}
}
\value{
Write this after testing.
}
\description{
Read vcf files and processes them / User correct assembly
}
\examples{

}
