% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv_Sort_Rm_VCF.R
\name{vcf2mut}
\alias{vcf2mut}
\title{Read vcf files and processes them / User correct assembly}
\usage{
vcf2mut(datapath, geno = "GT", assembly, five = FALSE)
}
\arguments{
\item{datapath}{The path leading to your vcf file.}

\item{geno}{An ID Format string in your vcf file indicating the index of the
genotype information.}

\item{assembly}{The path leading to your assembly file (.fa or .fa.gz).}

\item{five}{A boolean variable. A value of \code{TRUE} will lead to the function
scanning the input files for 5 bases mutation signatures as opposed to 3 bases
signatures. a value of \code{FALSE} causes the function to scan for 3 bases signatures.}
}
\value{
Write this after testing.
}
\description{
Read vcf files and processes them / User correct assembly
}
\examples{

}
